+++ runtime/src/lib.rs
@@ -162,17 +162,6 @@ pub fn native_version() -> NativeVersion {
 }
 
 frame_support::parameter_types! {
-	pub const MinimumPeriod: u64 = SLOT_DURATION / 2;
-}
-impl pallet_timestamp::Config for Runtime {
-	type MinimumPeriod = MinimumPeriod;
-	/// A timestamp: milliseconds since the unix epoch.
-	type Moment = u64;
-	type OnTimestampSet = ();
-	type WeightInfo = weights::pallet_timestamp::WeightInfo<Self>;
-}
-
-frame_support::parameter_types! {
 	pub const UncleGenerations: u32 = 0;
 }
 impl pallet_authorship::Config for Runtime {
