+++ runtime/src/pallets/mod.rs
@@ -1,2 +1,5 @@
 pub mod system;
 pub use system::*;
+
+pub mod timestamp;
+pub use timestamp::*;
