+++ runtime/src/pallets/timestamp.rs
@@ -1,0 +1,13 @@
+use crate::*;
+
+frame_support::parameter_types! {
+	pub const MinimumPeriod: u64 = SLOT_DURATION / 2;
+}
+
+impl pallet_timestamp::Config for Runtime {
+	type MinimumPeriod = MinimumPeriod;
+	/// A timestamp: milliseconds since the unix epoch.
+	type Moment = u64;
+	type OnTimestampSet = ();
+	type WeightInfo = weights::pallet_timestamp::WeightInfo<Self>;
+}
